+++ package.json
@@ -70,6 +70,7 @@
   "devDependencies": {
     "@babel/plugin-transform-react-display-name": "^7.10.4",
     "@babel/preset-typescript": "^7.10.4",
+    "@types/lodash": "^4.14.168",
     "@types/logrocket-react": "^3.0.0",
     "@types/react": "^16.9.55",
     "@types/react-dom": "^17.0.0",
